--- lua/gitlinks/init.lua
@@ -1,1 +1,0 @@
-require("gitlinks.core")
