+++ lua/gitportal/init.lua
@@ -1,0 +1,1 @@
+require("gitportal.core")
